[FEATURE] #14 - 경로 탐색 ViewModel 추가

+++ core/domain/src/main/java/com/eight_potato/domain/model/address/PoiModel.kt
@@ -4,6 +4,6 @@ package com.eight_potato.domain.model.address
  * 좌표 정보
  */
 data class PoiModel(
-    val lat: Float, // 위도
-    val lon: Float // 경도
+    val lat: Double, // 위도
+    val lon: Double // 경도
 )
